+++ src/GRA.Controllers/MissionControl/EventsController.cs
@@ -373,6 +373,8 @@ namespace GRA.Controllers.MissionControl
         [HttpPost]
         public async Task<IActionResult> Create(EventsDetailViewModel model)
         {
+            byte[] badgeBytes = null;
+
             var requireSecretCode = await GetSiteSettingBoolAsync(
                 SiteSettingKey.Events.RequireBadge);
 
@@ -409,10 +411,27 @@ namespace GRA.Controllers.MissionControl
                     ModelState.AddModelError("BadgemakerImage", "A badge is required.");
                 }
                 else if (model.BadgeUploadImage != null
-                    && (string.IsNullOrWhiteSpace(model.BadgeMakerImage) || !model.UseBadgeMaker)
-                    && (!ValidImageExtensions.Contains(Path.GetExtension(model.BadgeUploadImage.FileName).ToLower())))
+                    && (string.IsNullOrWhiteSpace(model.BadgeMakerImage) || !model.UseBadgeMaker))
                 {
-                    ModelState.AddModelError("BadgeUploadImage", $"Image must be one of the following types: {string.Join(", ", ValidImageExtensions)}");
+                    if (!ValidImageExtensions.Contains(
+                        Path.GetExtension(model.BadgeUploadImage.FileName).ToLowerInvariant()))
+                    {
+                        ModelState.AddModelError("BadgeUploadImage", $"Image must be one of the following types: {string.Join(", ", ValidImageExtensions)}");
+                    }
+
+                    try
+                    {
+                        using (var ms = new MemoryStream())
+                        {
+                            await model.BadgeUploadImage.CopyToAsync(ms);
+                            badgeBytes = ms.ToArray();
+                        }
+                        await _badgeService.ValidateBadgeImageAsync(badgeBytes);
+                    }
+                    catch (GraException gex)
+                    {
+                        ModelState.AddModelError("BadgeUploadImage", gex.Message);
+                    }
                 }
                 if (!UserHasPermission(Permission.IgnorePointLimits)
                     && model.MaxPointLimit.HasValue
@@ -484,7 +503,6 @@ namespace GRA.Controllers.MissionControl
                     int? triggerId = null;
                     if (model.IncludeSecretCode)
                     {
-                        byte[] badgeBytes;
                         string filename;
                         if (!string.IsNullOrWhiteSpace(model.BadgeMakerImage)
                             && (model.BadgeUploadImage == null || model.UseBadgeMaker))
@@ -495,11 +513,11 @@ namespace GRA.Controllers.MissionControl
                         }
                         else
                         {
-                            using (var fileStream = model.BadgeUploadImage.OpenReadStream())
+                            if (badgeBytes == null)
                             {
                                 using (var ms = new MemoryStream())
                                 {
-                                    fileStream.CopyTo(ms);
+                                    await model.BadgeUploadImage.CopyToAsync(ms);
                                     badgeBytes = ms.ToArray();
                                 }
                             }
@@ -575,6 +593,8 @@ namespace GRA.Controllers.MissionControl
         [HttpPost]
         public async Task<IActionResult> AddEditStreaming(EventsDetailViewModel model)
         {
+            byte[] badgeBytes = null;
+
             var requireSecretCode = await GetSiteSettingBoolAsync(
                 SiteSettingKey.Events.RequireBadge);
 
@@ -590,10 +610,27 @@ namespace GRA.Controllers.MissionControl
                     ModelState.AddModelError("BadgemakerImage", "A badge is required.");
                 }
                 else if (model.BadgeUploadImage != null
-                    && (string.IsNullOrWhiteSpace(model.BadgeMakerImage) || !model.UseBadgeMaker)
-                    && (!ValidImageExtensions.Contains(Path.GetExtension(model.BadgeUploadImage.FileName).ToLower())))
+                    && (string.IsNullOrWhiteSpace(model.BadgeMakerImage) || !model.UseBadgeMaker))
                 {
-                    ModelState.AddModelError("BadgeUploadImage", $"Image must be one of the following types: {string.Join(", ", ValidImageExtensions)}");
+                    if (!ValidImageExtensions.Contains(
+                        Path.GetExtension(model.BadgeUploadImage.FileName).ToLowerInvariant()))
+                    {
+                        ModelState.AddModelError("BadgeUploadImage", $"Image must be one of the following types: {string.Join(", ", ValidImageExtensions)}");
+                    }
+
+                    try
+                    {
+                        using (var ms = new MemoryStream())
+                        {
+                            await model.BadgeUploadImage.CopyToAsync(ms);
+                            badgeBytes = ms.ToArray();
+                        }
+                        await _badgeService.ValidateBadgeImageAsync(badgeBytes);
+                    }
+                    catch (GraException gex)
+                    {
+                        ModelState.AddModelError("BadgeUploadImage", gex.Message);
+                    }
                 }
             }
             else
@@ -615,7 +652,6 @@ namespace GRA.Controllers.MissionControl
                     int? triggerId = null;
                     if (model.IncludeSecretCode && !model.Editing)
                     {
-                        byte[] badgeBytes;
                         string filename;
                         if (!string.IsNullOrWhiteSpace(model.BadgeMakerImage)
                             && (model.BadgeUploadImage == null || model.UseBadgeMaker))
@@ -626,11 +662,11 @@ namespace GRA.Controllers.MissionControl
                         }
                         else
                         {
-                            using (var fileStream = model.BadgeUploadImage.OpenReadStream())
+                            if (badgeBytes == null)
                             {
                                 using (var ms = new MemoryStream())
                                 {
-                                    fileStream.CopyTo(ms);
+                                    await model.BadgeUploadImage.CopyToAsync(ms);
                                     badgeBytes = ms.ToArray();
                                 }
                             }
@@ -1283,7 +1319,8 @@ namespace GRA.Controllers.MissionControl
         {
             PageTitle = "Import Events";
             if (eventFileCsv == null
-                || !ValidCsvExtensions.Contains(Path.GetExtension(eventFileCsv.FileName).ToLower()))
+                || !ValidCsvExtensions
+                    .Contains(Path.GetExtension(eventFileCsv.FileName).ToLowerInvariant()))
             {
                 AlertDanger = "You must select a .csv file.";
                 ModelState.AddModelError("eventFileCsv", "You must select a .csv file.");