+++ src/GRA.Data.SqlServer/Migrations/20201016205549_add-event-challenge-favorites.cs
@@ -1,0 +1,66 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+namespace GRA.Data.SqlServer.Migrations
+{
+    public partial class addeventchallengefavorites : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<int>(
+                name: "EventId",
+                table: "UserLogs",
+                nullable: true);
+
+            migrationBuilder.AddColumn<int>(
+                name: "TriggerId",
+                table: "UserLogs",
+                nullable: true);
+
+            migrationBuilder.CreateTable(
+                name: "UserFavoriteEvents",
+                columns: table => new
+                {
+                    UserId = table.Column<int>(nullable: false),
+                    EventId = table.Column<int>(nullable: false),
+                    CreatedAt = table.Column<DateTime>(nullable: false),
+                    CreatedBy = table.Column<int>(nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_UserFavoriteEvents", x => new { x.UserId, x.EventId });
+                    table.ForeignKey(
+                        name: "FK_UserFavoriteEvents_Events_EventId",
+                        column: x => x.EventId,
+                        principalTable: "Events",
+                        principalColumn: "Id",
+                        onDelete: ReferentialAction.Restrict);
+                    table.ForeignKey(
+                        name: "FK_UserFavoriteEvents_Users_UserId",
+                        column: x => x.UserId,
+                        principalTable: "Users",
+                        principalColumn: "Id",
+                        onDelete: ReferentialAction.Restrict);
+                });
+
+            migrationBuilder.CreateIndex(
+                name: "IX_UserFavoriteEvents_EventId",
+                table: "UserFavoriteEvents",
+                column: "EventId");
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropTable(
+                name: "UserFavoriteEvents");
+
+            migrationBuilder.DropColumn(
+                name: "EventId",
+                table: "UserLogs");
+
+            migrationBuilder.DropColumn(
+                name: "TriggerId",
+                table: "UserLogs");
+        }
+    }
+}