+++ src/GRA.Data.SqlServer/Migrations/SqlServerContextModelSnapshot.cs
@@ -2733,6 +2733,23 @@ namespace GRA.Data.SqlServer.Migrations
                     b.ToTable("UserFavoriteChallenges");
                 });
 
+            modelBuilder.Entity("GRA.Data.Model.UserFavoriteEvent", b =>
+                {
+                    b.Property<int>("UserId");
+
+                    b.Property<int>("EventId");
+
+                    b.Property<DateTime>("CreatedAt");
+
+                    b.Property<int>("CreatedBy");
+
+                    b.HasKey("UserId", "EventId");
+
+                    b.HasIndex("EventId");
+
+                    b.ToTable("UserFavoriteEvents");
+                });
+
             modelBuilder.Entity("GRA.Data.Model.UserLog", b =>
                 {
                     b.Property<int>("Id")
@@ -2757,6 +2774,8 @@ namespace GRA.Data.SqlServer.Migrations
 
                     b.Property<string>("Description");
 
+                    b.Property<int?>("EventId");
+
                     b.Property<bool?>("HasBeenViewed");
 
                     b.Property<bool>("IsDeleted");
@@ -2764,6 +2783,8 @@ namespace GRA.Data.SqlServer.Migrations
                     b.Property<int?>("PointTranslationId");
 
                     b.Property<int>("PointsEarned");
+
+                    b.Property<int?>("TriggerId");
 
                     b.Property<int>("UserId");
 
@@ -3582,6 +3603,19 @@ namespace GRA.Data.SqlServer.Migrations
                     b.HasOne("GRA.Data.Model.Challenge", "Challenge")
                         .WithMany()
                         .HasForeignKey("ChallengeId")
+                        .OnDelete(DeleteBehavior.Restrict);
+
+                    b.HasOne("GRA.Data.Model.User", "User")
+                        .WithMany()
+                        .HasForeignKey("UserId")
+                        .OnDelete(DeleteBehavior.Restrict);
+                });
+
+            modelBuilder.Entity("GRA.Data.Model.UserFavoriteEvent", b =>
+                {
+                    b.HasOne("GRA.Data.Model.Event", "Event")
+                        .WithMany()
+                        .HasForeignKey("EventId")
                         .OnDelete(DeleteBehavior.Restrict);
 
                     b.HasOne("GRA.Data.Model.User", "User")