+++ src/GRA/ImageType.cs
@@ -1,0 +1,8 @@
+﻿namespace GRA
+{
+    public enum ImageType
+    {
+        Png,
+        Jpg
+    }
+}